change dtype Criteria.py to float

+++ Criteria_Based-RS.py
@@ -4,7 +4,7 @@ from sklearn.metrics.pairwise import *
 from scipy import sparse 
 import sys
 
-criteria = list(map(int, sys.argv[1:]))
+criteria = list(map(float, sys.argv[1:]))
 
 df = pd.read_csv(open("/home/yntn/Thesis/data/restaurant-foody-ver2.csv",'r'))
 go = df.values[:,-5:]
